+++ src/hooks.ts
@@ -2,8 +2,9 @@ import { After, AfterStep, Before, BeforeStep, Status } from '@cucumber/cucumber
 import defaultTimeouts from './defaultTimeouts';
 import { Browser, remote } from 'webdriverio';
 import { po } from '@qavajs/po';
-import { ScreenshotEvent } from './screenshotEvent';
+import { ScreenshotEvent, SnapshotEvent } from './events';
 import { equalOrIncludes } from './utils';
+import getSnapshot from './client_script/snapshot';
 
 declare global {
     var browser: Browser;
@@ -32,7 +33,7 @@ Before(async function () {
 });
 
 BeforeStep(async function () {
-    const isBeforeStepScreenshot = equalOrIncludes(config.screenshot, ScreenshotEvent.BEFORE_STEP);
+    const isBeforeStepScreenshot = equalOrIncludes(config.driverConfig.screenshot ?? config.screenshot, ScreenshotEvent.BEFORE_STEP);
     if (isBeforeStepScreenshot) {
         try {
             this.attach(await browser.takeScreenshot(), 'base64:image/png');
@@ -40,17 +41,38 @@ BeforeStep(async function () {
             console.warn(err)
         }
     }
+    const isBeforeStepSnapshot = equalOrIncludes(config.driverConfig.snapshot, SnapshotEvent.BEFORE_STEP);
+    if (isBeforeStepSnapshot) {
+        try {
+            this.attach(Buffer.from(await browser.executeAsync(getSnapshot)).toString('base64'), 'text/html');
+        } catch (err) {
+            console.warn(err)
+        }
+    }
 });
 
 AfterStep(async function (step) {
-    const isAfterStepScreenshot = equalOrIncludes(config.screenshot, ScreenshotEvent.AFTER_STEP);
-    const isOnFailScreenshot = equalOrIncludes(config.screenshot, ScreenshotEvent.ON_FAIL);
+    const isAfterStepScreenshot = equalOrIncludes(config.driverConfig.screenshot ?? config.screenshot, ScreenshotEvent.AFTER_STEP);
+    const isOnFailScreenshot = equalOrIncludes(config.driverConfig.screenshot ?? config.screenshot, ScreenshotEvent.ON_FAIL)
     try {
         if (
             (isOnFailScreenshot && step.result.status === Status.FAILED) ||
             isAfterStepScreenshot
         ) {
             this.attach(await browser.takeScreenshot(), 'base64:image/png');
+        }
+    } catch (err) {
+        console.warn(err)
+    }
+
+    const isAfterStepSnapshot = equalOrIncludes(config.driverConfig.snapshot, SnapshotEvent.AFTER_STEP);
+    const isOnFailSnapshot = equalOrIncludes(config.driverConfig.snapshot, SnapshotEvent.ON_FAIL);
+    try {
+        if (
+            (isOnFailSnapshot && step.result.status === Status.FAILED) ||
+            isAfterStepSnapshot
+        ) {
+            this.attach(Buffer.from(await browser.executeAsync(getSnapshot)).toString('base64'), 'text/html');
         }
     } catch (err) {
         console.warn(err)